requirements.txt
Flask>=3.0.3
gunicorn>=22.0.0
python-docx>=1.1.2
WTForms>=3.1.2
email-validator>=2.2.0
